Changes requestd by reviewer
