+++ integration/src/test/java/org/kie/hacep/RemoteKieSessionImplTest.java
@@ -21,7 +21,6 @@ import java.util.concurrent.TimeUnit;
 
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.kie.hacep.core.Bootstrap;
 import org.kie.hacep.core.infra.election.State;
@@ -40,7 +39,7 @@ public class RemoteKieSessionImplTest {
     @Before
     public void setUp() throws Exception {
         topicsConfig = TopicsConfig.getDefaultTopicsConfig();
-        envConfig = TestUtils.getEnvConfig();
+        envConfig = KafkaUtilTest.getEnvConfig();
         kafkaServerTest = new KafkaUtilTest();
         kafkaServerTest.startServer();
         kafkaServerTest.createTopics(topicsConfig.getEventsTopicName(),
