--- integration/src/test/java/org/kie/hacep/TestUtils.java
@@ -1,19 +1,0 @@
-package org.kie.hacep;
-
-import org.kie.remote.CommonConfig;
-
-public class TestUtils {
-
-    public static EnvConfig getEnvConfig() {
-        return EnvConfig.anEnvConfig().
-                withNamespace(CommonConfig.DEFAULT_NAMESPACE).
-                withControlTopicName(Config.DEFAULT_CONTROL_TOPIC).
-                withEventsTopicName(CommonConfig.DEFAULT_EVENTS_TOPIC).
-                withSnapshotTopicName(Config.DEFAULT_SNAPSHOT_TOPIC).
-                withKieSessionInfosTopicName(CommonConfig.DEFAULT_KIE_SESSION_INFOS_TOPIC).
-                withPrinterType(PrinterKafkaImpl.class.getName()).
-                withPollTimeout("10").
-                isUnderTest(Boolean.TRUE.toString()).build();
-    }
-
-}
